Convert stats presenter to frame-stepped StatsState
Replace the blocking presentStats() loop with a StatsState that
renders one frame per update/draw cycle. GamePlayState now pushes
StatsState via replaceTop() when the controller signals game-over,
instead of the controller calling the blocking function directly.

This eliminates the last blocking gfx.process() call in the
application, making the entire loop fully frame-stepped.
