+++ src/game/controller/replayController.cpp
@@ -1,7 +1,6 @@
 #include "replayController.hpp"
 
 #include "../game.hpp"
-#include "stats_presenter.hpp"
 #include "../spectatorviewport.hpp"
 #include "../viewport.hpp"
 #include "../sfx.hpp"
@@ -126,7 +125,6 @@ bool ReplayController::process()
 		if (state == StateGameEnded)
 		{
 			game->statsRecorder->finish(*game);
-			presentStats(static_cast<NormalStatsRecorder&>(*game->statsRecorder), *game);
 		}
 		return false;
 	}
@@ -134,7 +132,6 @@ bool ReplayController::process()
 	if (!replay.get() && state == StateGame)
 	{
 		game->statsRecorder->finish(*game);
-		presentStats(static_cast<NormalStatsRecorder&>(*game->statsRecorder), *game);
 		return false;
 	}
