--- src/game/controller/stats_presenter.hpp
@@ -1,5 +1,0 @@
-#pragma once
-
-#include "../stats_recorder.hpp"
-
-void presentStats(NormalStatsRecorder& recorder, Game& game);
