+++ src/game/statsState.hpp
@@ -1,0 +1,34 @@
+#pragma once
+
+#include "state.hpp"
+#include "stats_recorder.hpp"
+#include "gfx/bitmap.hpp"
+#include <vector>
+
+struct Game;
+
+// Frame-stepped post-game statistics display.
+// Replaces the blocking presentStats() loop.
+struct StatsState : AppState
+{
+	StatsState(NormalStatsRecorder& recorder, Game& game);
+
+	void enter() override;
+	void handleEvent(SDL_Event& ev) override;
+	bool update() override;
+	void draw() override;
+	bool wantsMenuFlip() const override { return false; }
+
+private:
+	NormalStatsRecorder& recorder_;
+	Game& game_;
+
+	Bitmap bg_;
+	double offset_ = 0, destOffset_ = 0;
+	double pane_ = 0, destPane_ = 0;
+
+	std::vector<WeaponStats> combinedWeaponStats_;
+	std::vector<WeaponStats> weaponStats_[2];
+	std::vector<double> wormDamages_[2];
+	std::vector<double> wormTotalHpDiff_;
+};
